@@ -3,7 +3,7 @@
     <groupId>net.whydah.service</groupId>
     <artifactId>Whydah-OAuth2Service</artifactId>
     <name>Whydah-OAuth2Service</name>
-    <version>2.27.1-SNAPSHOT</version>
+    <version>2.30.0-SNAPSHOT</version>
     <url>https://wiki.cantara.no/display/whydah/OAuth2Service</url>
     <parent>
         <groupId>no.cantara.emi</groupId>
@@ -11,16 +11,18 @@
         <version>8</version>
     </parent>
     <properties>
+        <jdk.version>21</jdk.version>
         <service.name>Whydah-OAuth2Service</service.name>
-        <whydah-admin-sdk-version>2.27.3</whydah-admin-sdk-version>
+        <whydah-admin-sdk-version>2.30.0</whydah-admin-sdk-version>
         <jersey.version>2.46</jersey.version>
         <jetty.version>9.4.57.v20241219</jetty.version>
         <spring.version>5.3.39</spring.version>
         <slf4j.version>2.0.17</slf4j.version>
         <hystrix.version>1.5.18</hystrix.version>
-        <java.version>11</java.version>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
+        <maven.compiler.release>21</maven.compiler.release>
+        <java.version>21</java.version>
         <maven.javadoc.skip>true</maven.javadoc.skip>
     </properties>
     <scm>
@@ -338,6 +340,11 @@
                     <artifactId>jackson-databind</artifactId>
                 </exclusion>
             </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>jakarta.inject</groupId>
+            <artifactId>jakarta.inject-api</artifactId>
+            <version>1.0.3</version>
         </dependency>
         <!-- https://mvnrepository.com/artifact/com.nimbusds/nimbus-jose-jwt -->
         <dependency>
