@@ -1,13 +1,12 @@
 package net.whydah.service.errorhandling;
 
-import java.io.PrintWriter;
-import java.io.StringWriter;
-
 import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
+import java.io.PrintWriter;
+import java.io.StringWriter;
 
 @Provider
 public class GenericExceptionMapper implements ExceptionMapper<Throwable> {
@@ -34,8 +33,8 @@ public class GenericExceptionMapper implements ExceptionMapper<Throwable> {
 	}
 
 	private void setHttpStatus(Throwable ex, ErrorMessage errorMessage) {
-		if(ex instanceof WebApplicationException ) { 
-			errorMessage.setStatus(((WebApplicationException)ex).getResponse().getStatus());
+		if (ex instanceof WebApplicationException exception) {
+			errorMessage.setStatus(exception.getResponse().getStatus());
 		} else {
 			errorMessage.setStatus(Response.Status.INTERNAL_SERVER_ERROR.getStatusCode()); //defaults to internal server error 500
 		}
